app/telegram_bot/requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
redis==5.0.8
httpx==0.27.2
pydantic==2.8.2
Pillow==10.4.0
pytesseract==0.3.13
